refactor: add Compilation and Event public and admin controllers

+++ ewm-main-service/main-server/src/main/java/ru/practicum/ewm/main/service/category/service/CategoryServiceImpl.java
@@ -9,8 +9,11 @@ import ru.practicum.ewm.main.service.category.dto.CategoryDto;
 import ru.practicum.ewm.main.service.category.mapper.CategoryMapper;
 import ru.practicum.ewm.main.service.category.model.Category;
 import ru.practicum.ewm.main.service.category.repository.CategoryRepository;
+import ru.practicum.ewm.main.service.event.model.Event;
+import ru.practicum.ewm.main.service.event.repository.EventRepository;
+import ru.practicum.ewm.main.service.exception.ConflictException;
 import ru.practicum.ewm.main.service.exception.EntityNotFoundException;
-import ru.practicum.ewm.main.service.exception.ValidationException;
+import ru.practicum.ewm.main.service.exception.CustomValidationException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -21,11 +24,19 @@ public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Override
     public void delete(Long catId) {
         getOrThrow(catId);
-        categoryRepository.deleteById(catId);
+
+        List<Event> events = eventRepository.findAllByCategoryId(catId);
+
+        if (events.isEmpty()) {
+            categoryRepository.deleteById(catId);
+        } else {
+            throw new ConflictException("The category is not empty");
+        }
     }
 
     @Override
@@ -35,7 +46,7 @@ public class CategoryServiceImpl implements CategoryService {
             Category categorySaved = categoryRepository.save(category);
             return categoryMapper.categoryToDto(categorySaved);
         } catch (DataIntegrityViolationException e) {
-            throw new ValidationException("Category with specified name is already exists.");
+            throw new CustomValidationException("Category with specified name is already exists.");
         }
     }
